fix name menu_comands -> menu_commands
